Исправить schema

+++ stats/analyzer/src/main/resources/schema.sql
@@ -1,7 +1,7 @@
 CREATE TABLE IF NOT EXISTS event_similarities (
     first_event BIGINT,
     second_event BIGINT,
-    score DOUBLE,
+    score DOUBLE PRECISION,
     timestamp_action TIMESTAMP WITHOUT TIME ZONE,
     PRIMARY KEY(first_event, second_event)
 );
@@ -9,6 +9,6 @@ CREATE TABLE IF NOT EXISTS event_similarities (
 CREATE TABLE IF NOT EXISTS user_actions (
     user_id BIGINT,
     event_id BIGINT,
-    user_score DOUBLE,
+    user_score DOUBLE PRECISION,
     PRIMARY KEY(user_id, event_id)
 );
